move cmd/util to pkg/util/cli
