+++ cmd/user-ssh-keys-agent/main.go
@@ -27,9 +27,9 @@ import (
 	"github.com/go-logr/zapr"
 	"go.uber.org/zap"
 
-	cmdutil "k8c.io/kubermatic/v2/cmd/util"
 	usersshkeys "k8c.io/kubermatic/v2/pkg/controller/usersshkeysagent"
 	kubermaticlog "k8c.io/kubermatic/v2/pkg/log"
+	"k8c.io/kubermatic/v2/pkg/util/cli"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
@@ -45,7 +45,7 @@ func main() {
 	rawLog := kubermaticlog.New(logOpts.Debug, logOpts.Format)
 	log := rawLog.Sugar()
 
-	cmdutil.Hello(log, "User SSH-Key Agent", logOpts.Debug)
+	cli.Hello(log, "User SSH-Key Agent", logOpts.Debug)
 
 	cfg, err := config.GetConfig()
 	if err != nil {
